requirements.txt
requests==2.28.1
azure-storage-blob==12.14.1
azure-identity==1.13.0
azure-mgmt-resource==21.1.0
azure-mgmt-storage==19.0.0
python-dotenv